+++ index.js
@@ -1,4 +1,6 @@
 exports.Base = require('./lib/base');
 
+exports.Inflections = require('./lib/inflections');
+