+++ tests/phones.js
@@ -9,3 +9,6 @@ function Phone (){
 }
 Phone.super_();
 Phone.belongs_to('user')
+
+Phone.validate_presence_of('number')
+Phone.validate_length_of('number', {minimum: 10, maximum: 10})